adding sleep data and fixing time duration calculation

+++ app/src/main/java/uk/ac/aber/dcs/cs39440/myvitalife/model/Sleep.kt
@@ -1,7 +1,7 @@
 package uk.ac.aber.dcs.cs39440.myvitalife.model
 
 data class Sleep(
-    val score: Int = 0,
+    val score: Float = 0f,
     val start: String = "",
     val end: String = "",
     val duration: String = "",
